agora funcionando tudo filtro sharpen está estranho

+++ blur.m
@@ -1,7 +1,7 @@
 function blur(f, outputName)
 
   w = [1, 2, 1; 2, 4, 2; 1, 2, 1]/16;
-  g = convol(f, w);
+  g = uint8(convol(f, w));
   imwrite(g, outputName);
 
 endfunction
